Time Code Only Update
- Removed Sun and Moon calculations and related variables/funtions
- Renamed to SkyPlugin, Blueprint renamed to TimeManager
- IWYU compliant
