--- Plugins/TimeManager/TimeManager.uplugin
@@ -1,23 +1,0 @@
-{
-	"FileVersion": 3,
-	"Version": 1,
-	"VersionName": "1.0",
-	"FriendlyName": "TimeManager",
-	"Description": "Test",
-	"Category": "Other",
-	"CreatedBy": "DotCam",
-	"CreatedByURL": "Test",
-	"DocsURL": "",
-	"MarketplaceURL": "",
-	"SupportURL": "",
-	"CanContainContent": true,
-	"IsBetaVersion": false,
-	"Installed": false,
-	"Modules": [
-    {
-      "Name": "TimeManager",
-      "Type": "Developer",
-      "LoadingPhase": "PostEngineInit"
-    }
-	]
-}
