+++ Plugins/TimePlugin/Source/TimePlugin/Classes/TimeManager.h
@@ -23,7 +23,7 @@
 #include "CoreMinimal.h"
 #include "GameFramework/Actor.h"
 #include "TimeDateStruct.h"
-#include "Blueprint.generated.h"
+#include "TimeManager.generated.h"
 
 
 //An actor based calendar system for tracking date + time, and Sun/Moon rotation/phase.
@@ -46,6 +46,10 @@ public:
 	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "Default")
 		float Longitude = 0.0f;
 
+	// The value of the local Standard Time Meridian (15deg intervals)
+	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
+		int32 LSTM = 0;
+
 	// The number of hours offset from UTC for the local location (value in the range of -12 to +12 hours from UTC)	
 	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "Default")
 		int32 OffsetUTC = 0;
@@ -66,89 +70,8 @@ public:
 	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "Default")
 		float TimeScaleMultiplier = 1.0f;
 
-	// The current Local Solar Time (in minutes)
-	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
-		float SolarTime = 0.0f;
-
-	// The current Local Clock Time value (in minutes)
-	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
-		float LocalClockTime = 0.0f;
-
-	// The current Time Correction factor
-	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
-		float TimeCorrection = 0.0f;
-
-	// The value of the local Standard Time Meridian (15deg intervals)
-	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
-		int32 LSTM = 0;
-
 	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
 		int32 DayOfYear = 0;
-
-	// THe current Equation of Time value
-	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
-		float EoT = 0.0f;
-
-	// The current Altitude Angle of the sun
-	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
-		float SolarAltAngle = 0.0f;
-
-	// The current Declination angle of the sun
-	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
-		float SolarDeclination = 0.0f;
-
-	// The Azimuth angle of the sun
-	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
-		float SolarAzimuth = 0.0f;
-
-	// The current Hour Angle of the sun
-	UPROPERTY(BlueprintReadOnly, Category = "Sun Debug")
-		float SolarHRA = 0.0f;
-
-	// The current Sidereal Time value
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float SiderealTime = 0.0f;
-
-	// The current Altitude Angle of the moon
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float LunarAltAngle = 0.0f;
-
-	// The current Hour Angle of the moon
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float LunarHRA = 0.0f;
-
-	// The current Declination angle of the moon
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float LunarDeclination = 0.0f;
-
-	// The current Azimuth angle of the moon
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float LunarAzimuth = 0.0f;
-
-	// The current Right Ascension angle for the moon
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float LunarRightAsc = 0.0f;
-
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float LunarElapsedDays = 0.0f;
-
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float EcLongitude = 0.0f;
-
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float EcLatitude = 0.0f;
-
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float EcDistance = 0.0f;
-
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float PartL = 0.0f;
-
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float PartM = 0.0f;
-
-	UPROPERTY(BlueprintReadOnly, Category = "Lunar Debug")
-		float PartF = 0.0f;
 
 	// -------------------
 	// PUBLIC FUNCTIONS
@@ -256,35 +179,6 @@ public:
 		void IncrementTime(float deltaSeconds);
 
 
-	/**
-	* @Name: CalculateSunAngle
-	* @Description: Calculates the sun angle rotator from the current time.
-	*
-	* @return: FRotator - The sun rotation value for the current time.
-	*/
-	UFUNCTION(BlueprintCallable, Category = "TimeManager")
-		FRotator CalculateSunAngle();
-
-
-	/**
-	* @Name: CalculateMoonAngle
-	* @Description: Calculates the moon angle rotator from the current time.
-	*
-	* @return: FRotator - The moon rotation value for the current time.
-	*/
-	UFUNCTION(BlueprintCallable, Category = "TimeManager")
-		FRotator CalculateMoonAngle();
-
-
-	/**
-	* @Name: CalculateMoonPhase
-	* @Description: Calculates the moon phase for the current time and date.
-	*
-	* @return: FRotator - The moon rotation value for the current time.
-	*/
-	UFUNCTION(BlueprintCallable, Category = "TimeManager")
-		float CalculateMoonPhase();
-
 
 private:
 
@@ -302,57 +196,6 @@ private:
 	double JD1900 = 2415020.0;
 
 	double ElapsedJD1900 = 0.0;
-
-	// Obliquity of the Ecliptic (as of 2000/01/01 - approximation, but fairly accurate)
-	double EcObliquity = 23.4397;
-
-	/* --- Utility Functions --- */
-
-	// Float versions
-
-	// Performs FMath::Sin(input) using degrees
-	float SinD(float input);
-
-	// Performs FMath::Asin(input) using degrees
-	float ASinD(float input);
-
-	// Performs FMath::Cos(input) using degrees
-	float CosD(float input);
-
-	// Performs FMath::Acos(input) using degrees
-	float ACosD(float input);
-
-	// Performs FMath::Tan(input) using degrees
-	float TanD(float input);
-
-	// Performs FMath::Atan(input) using degrees
-	float ATanD(float input);
-
-	// Performs FMath::Atan2(A, B) using degrees
-	float ATan2D(float A, float B);
-
-	// Double versions
-
-	// Performs FMath::Sin(input) using degrees
-	double SinD(double input);
-
-	// Performs FMath::Asin(input) using degrees
-	double ASinD(double input);
-
-	// Performs FMath::Cos(input) using degrees
-	double CosD(double input);
-
-	// Performs FMath::Acos(input) using degrees
-	double ACosD(double input);
-
-	// Performs FMath::Tan(input) using degrees
-	double TanD(double input);
-
-	// Performs FMath::Atan(input) using degrees
-	double ATanD(double input);
-
-	// Performs FMath::Atan2(A, B) using degrees
-	double ATan2D(double A, double B);
 
 	FTimeDateStruct ConvertToTimeDate(FDateTime dt);
