+++ Plugins/TimePlugin/Source/TimePlugin/Classes/TimePluginPrivatePCH.h
@@ -1,0 +1,5 @@
+// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
+#pragma once
+
+#include "CoreMinimal.h"
+#include "EngineMinimal.h"
