+++ Plugins/TimePlugin/Source/TimePlugin/Private/TimePlugin.cpp
@@ -1,29 +1,28 @@
 // Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
-#include "IModule.h"
+#include "TimePlugin.h"
 #include "EngineUtils.h"
 
 
-#define LOCTEXT_NAMESPACE "FTimeManagerModule"
 
-void FTimeManagerModule::StartupModule()
+void FTimePlugin::StartupModule()
 {
 	// This code will execute after your module is loaded into memory; the exact timing is specified in the .uplugin file per-module
 	//Create our delegate type
 	FWorldDelegates::FWorldInitializationEvent::FDelegate OnWorldCreatedDelegate;
 	//Declare which function we want to bind to
-	OnWorldCreatedDelegate = FWorldDelegates::FWorldInitializationEvent::FDelegate::CreateRaw(this, &FTimeManagerModule::OnWorldCreated);
+	OnWorldCreatedDelegate = FWorldDelegates::FWorldInitializationEvent::FDelegate::CreateRaw(this, &FTimePlugin::OnWorldCreated);
 	//Declare which event we want to bind to
 	FDelegateHandle OnWorldCreatedDelegateHandle = FWorldDelegates::OnPostWorldInitialization.Add(OnWorldCreatedDelegate);
 }
 
-void FTimeManagerModule::ShutdownModule()
+void FTimePlugin::ShutdownModule()
 {
 	// This function may be called during shutdown to clean up your module.  For modules that support dynamic reloading,
 	// we call this function before unloading the module.
 }
 
-void FTimeManagerModule::OnWorldCreated(UWorld* World, const UWorld::InitializationValues IVS)
+void FTimePlugin::OnWorldCreated(UWorld* World, const UWorld::InitializationValues IVS)
 {
 	//If we already have a TimeManagerActor do not spawn another one
 	//Just store it as the current TimeManagerActor for other plugins to use
@@ -38,7 +37,6 @@ void FTimeManagerModule::OnWorldCreated(UWorld* World, const UWorld::Initializat
 	TimeManagerActor = World->SpawnActor<ATimeManager>(ATimeManager::StaticClass(), location, rotate, SpawnInfo);
 }
 
-#undef LOCTEXT_NAMESPACE
 	
-IMPLEMENT_MODULE(FTimeManagerModule, TimeManager)
+IMPLEMENT_MODULE(FTimePlugin, TimePlugin)
