+++ Plugins/TimePlugin/Source/TimePlugin/TimePlugin.Build.cs
@@ -1,0 +1,34 @@
+// Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
+
+using UnrealBuildTool;
+using System.IO;
+
+public class TimePlugin : ModuleRules
+{
+    private string ModulePath
+    {
+        get { return Path.GetDirectoryName(ModuleDirectory); }
+    }
+
+    private string ThirdPartyPath
+    {
+        get { return Path.GetFullPath(Path.Combine(ModulePath, "../../ThirdParty/")); }
+    }
+
+    public TimePlugin(ReadOnlyTargetRules Target) : base(Target)
+    {
+        PrivateIncludePaths.AddRange(new string[] { "TimePlugin/Private" });
+
+        // Make sure UBT reminds us of how to keep the project IWYU compliant
+        bEnforceIWYU = true;
+
+        //Enable IWYU but keep our PrivatePCH in use
+        PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
+
+        //Our PrivatePCH that we want to globally #include
+        PrivatePCHHeaderFile = "Classes/TimePluginPrivatePCH.h";
+
+        PublicDependencyModuleNames.AddRange(new string[] { "Engine", "Core", "CoreUObject", "InputCore", "RHI", "RenderCore" });
+    }
+}
+
